Extract AppShellView from AppFeature reducer
