+++ Sources/Features/AppFeature/AppFeature.swift
@@ -1,5 +1,4 @@
 import ComposableArchitecture
-import SwiftUI
 
 enum AppTab: Hashable {
     case chart
@@ -104,116 +103,4 @@ struct AppFeature: Reducer {
     private func mapStatus(from payload: DailyPricingSnapshotPayload, whenNotEmpty status: RootStatus) -> RootStatus {
         payload.hourlyPrices.isEmpty ? .empty : status
     }
-}
-
-struct AppShellView: View {
-    let store: StoreOf<AppFeature>
-
-    var body: some View {
-        ZStack {
-            backgroundView
-            tabView
-            statusBannerOverlay
-        }
-        .onAppear {
-            store.send(.onAppear)
-        }
-    }
-
-    private var backgroundView: some View {
-        Color(.systemBackground)
-            .ignoresSafeArea()
-    }
-
-    private var statusBannerOverlay: some View {
-        RootStatusBanner(
-            onRetry: { store.send(.retryTapped) },
-            status: store.rootStatus
-        )
-        .padding(.horizontal, 16)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .center)
-        .accessibilityIdentifier("appRootStatusBanner")
-    }
-
-    private var tabSelection: Binding<AppTab> {
-        Binding(
-            get: { store.selectedTab },
-            set: { store.send(.selectedTabChanged($0)) }
-        )
-    }
-
-    private var tabView: some View {
-        TabView(selection: tabSelection) {
-            PricesView()
-                .tabItem {
-                    Label(AppTab.prices.title, systemImage: AppTab.prices.systemImage)
-                }
-                .tag(AppTab.prices)
-
-            ChartView()
-                .tabItem {
-                    Label(AppTab.chart.title, systemImage: AppTab.chart.systemImage)
-                }
-                .tag(AppTab.chart)
-
-            SettingsView()
-                .tabItem {
-                    Label(AppTab.settings.title, systemImage: AppTab.settings.systemImage)
-                }
-                .tag(AppTab.settings)
-        }
-        .accessibilityIdentifier("appTabView")
-    }
-}
-
-#Preview("App shell - loading") {
-    AppShellView(
-        store: Store(
-            initialState: AppFeature.State(
-                rootStatus: .loading,
-                selectedTab: .prices
-            )
-        ) {
-            AppFeature()
-        }
-    )
-}
-
-#Preview("App shell - error") {
-    AppShellView(
-        store: Store(
-            initialState: AppFeature.State(
-                rootStatus: .error,
-                selectedTab: .prices
-            )
-        ) {
-            AppFeature()
-        }
-    )
-}
-
-#Preview("App shell - chart") {
-    AppShellView(
-        store: Store(
-            initialState: AppFeature.State(
-                rootStatus: .content,
-                selectedTab: .chart
-            )
-        ) {
-            AppFeature()
-        }
-    )
-}
-
-#Preview("App shell - settings") {
-    AppShellView(
-        store: Store(
-            initialState: AppFeature.State(
-                rootStatus: .cached,
-                selectedTab: .settings
-            )
-        ) {
-            AppFeature()
-        }
-    )
 }
